backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.15
openai==1.3.0
python-multipart==0.0.6
pydantic==2.8.2
sentence-transformers==2.2.2
pypdf==3.17.0
python-docx==1.1.0
beautifulsoup4==4.12.2
email-validator==2.1.0
sqlalchemy==2.0.23
bcrypt==4.0.1
python-jose==3.3.0
aiofiles==23.2.1
google-generativeai==0.3.0
python-dotenv==1.0.0